Refactoring: move updateQuality(Item item) method from the Inn class to the Item class + update unit tests
