+++ src/main/java/org/agilegrenoble/objectcalisthenics/Item.java
@@ -30,4 +30,62 @@ public class Item {
 	public int getQuality() {
 		return quality;
 	}
+
+	/**
+	 * Update quality for the item
+	 * 
+	 */
+	public void updateQuality() {
+		if (!getName().equals("Aged Brie")
+				&& !getName().equals(
+						"Backstage passes to a TAFKAL80ETC concert")) {
+			if (getQuality() > 0) {
+				if (!getName().equals("Sulfuras, Hand of Ragnaros")) {
+					setQuality(getQuality() - 1);
+				}
+			}
+		} else {
+			if (getQuality() < 50) {
+				setQuality(getQuality() + 1);
+
+				if (getName().equals(
+						"Backstage passes to a TAFKAL80ETC concert")) {
+					if (getSellIn() < 11) {
+						if (getQuality() < 50) {
+							setQuality(getQuality() + 1);
+						}
+					}
+
+					if (getSellIn() < 6) {
+						if (getQuality() < 50) {
+							setQuality(getQuality() + 1);
+						}
+					}
+				}
+			}
+		}
+
+		if (!getName().equals("Sulfuras, Hand of Ragnaros")) {
+			setSellIn(getSellIn() - 1);
+		}
+
+		if (getSellIn() < 0) {
+			if (!getName().equals("Aged Brie")) {
+				if (!getName().equals(
+						"Backstage passes to a TAFKAL80ETC concert")) {
+					if (getQuality() > 0) {
+						if (!getName().equals("Sulfuras, Hand of Ragnaros")) {
+							setQuality(getQuality() - 1);
+						}
+					}
+				} else {
+					setQuality(getQuality() - getQuality());
+				}
+			} else {
+				if (getQuality() < 50) {
+					setQuality(getQuality() + 1);
+				}
+			}
+		}
+	}
 }
